[wrangler] Discard a pending assets refresh on teardown and config reload
The debounced bundle refresh armed by the assets watcher was never cancelled, so
a pending refresh could dispatch `bundleComplete` into a torn-down bus, or report
a bundle against a config that a reload had just replaced. Its timer could also
delay the process exiting by up to the debounce delay.

Store it on a field like `#debouncedCustomBuild` and cancel it once the old
watcher has closed — both when re-watching for a new config and in `teardown()`
— plus a `tearingDown` guard for events delivered while the watcher is closing.

Renames the EMFILE test file, which already mocks chokidar, to cover assets
watcher behaviour generally. Driving the watcher through the mock makes the new
tests deterministic: both fail without the cancellation.
